+++ app/controllers/api/v1/project_columns_controller.rb
@@ -1,6 +1,4 @@
 class Api::V1::ProjectColumnsController < Api::V1::ApiController
-  before_action :check_if_user_activated
-
   def show
     skip_authorization
     project = current_user.projects.find(params[:project_id])