+++ app/javascript/components/api/v1/AuthAPI.ts
@@ -18,10 +18,12 @@ export interface IUserProfile {
     is_superadmin: boolean;
 }
 
-export interface ILoginResponse {
-    data?: IUserProfile;
-    errors?: any[];
-}
+export type ILoginResponse =
+    | {
+          data?: IUserProfile;
+          errors?: any[];
+      }
+    | { error: true; error_type: "USER_IS_DEACTIVATED" };
 
 const AuthAPI = {
     signup: async (username: string, email: string, password: string, passwordConfirmation: string): Promise<any> => {